feat(examples): add CLI dev script for interactive module testing
