+++ apps/examples/src/seller/getActiveShops.ts
@@ -7,7 +7,7 @@ const sdk = new TikTokShopSDK({
     appSecret: process.env.TIKTOK_APP_SECRET!,
 });
 
-async function main(){
+export async function main(){
 
     try {
